fix(hr): the exported detrend fit needs 64 bits, not 32
a_q16 is the window's DC level shifted left 16, and this sensor's raw counts
are 24-bit, so the intercept sits around 1e11. The exported copy was int32 and
wrapped: across the first six hours of real dumps fit_a spanned the entire
int32 range and 49% of the values came back negative, for a signal that is
positive by construction.

The measurement itself was never wrong -- detrend_into_work computes in int64
and only the exported copy was narrowed -- but the dump could not be inverted
back to raw counts, which was the entire point of shipping it.

The samples were fine: over 330 windows, raw >> 3 matched the paired int8 dump
on every single value, which is what localised this to the two fit fields.

Wire header 15 -> 23 bytes, 279 B per chunk, still one frame.

+++ example/get-started/dualcore/src/modules/communicate/communicate_task.h
@@ -79,7 +79,7 @@ bool commu_send_sleep_diag(uint32_t ts, uint16_t score, uint8_t hr,
 /* One chunk of the SAME window at full precision, invertible back to raw
    sensor counts (@ref KEY_HR_WINDOW_RAW). count <= 128; the caller sends two
    chunks per window and repeats the fit in both. */
-bool commu_send_hr_window_raw(uint32_t ts, int32_t fit_a_q16, int32_t fit_b_q16,
+bool commu_send_hr_window_raw(uint32_t ts, int64_t fit_a_q16, int64_t fit_b_q16,
                               uint8_t shift, uint16_t first_index,
                               uint16_t count, const int16_t *win);
